add test folder to peerset
